requirements.txt
setuptools
dataclass_csv
google-cloud-spanner
Faker
